First publishable sample
Creates the main entrypoint `index.js` file that exports the echo
function that we want to ship in this package.

Also creates a `.npmrc` file with a custom registry so that we don't
have to publish to the offial NPM registry for a demo package for the
purposes of this lesson.